Changed css for navbar so search hides on small screens and updated css for post listing on homepage
