+++ src/main/java/com/stevedevblog/app/DevBlogApplication.java
@@ -1,0 +1,13 @@
+package com.stevedevblog.app;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DevBlogApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(DevBlogApplication.class, args);
+	}
+
+}
