+++ src/main/java/com/stevedevblog/app/domain/PostCategory.java
@@ -1,0 +1,18 @@
+package com.stevedevblog.app.domain;
+
+public enum PostCategory {
+    BUILD_LOG("Build Log"),
+    TESTING("Testing"),
+    DESIGN_PATTERNS("Design Patterns");
+
+    private final String value;
+
+    PostCategory(String value) {
+
+        this.value = value;
+    }
+
+    public String getValue() {
+        return value;
+    }
+}
