+++ src/main/java/com/stevedevblog/app/security/SecurityConfigurationService.java
@@ -1,0 +1,53 @@
+package com.stevedevblog.app.security;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.provisioning.InMemoryUserDetailsManager;
+
+import static com.stevedevblog.app.security.UserRole.ADMIN;
+
+@Configuration
+@EnableWebSecurity
+public class SecurityConfigurationService extends WebSecurityConfigurerAdapter {
+
+    private final PasswordEncoder passwordEncoder;
+
+    @Autowired
+    public SecurityConfigurationService(PasswordEncoder passwordEncoder) {
+        this.passwordEncoder = passwordEncoder;
+    }
+
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        http
+                .csrf().disable()
+                .authorizeRequests()
+                .antMatchers("/", "/error", "/css/*", "/js/*", "/icons/*", "/blog-post/*").permitAll()
+                .antMatchers("/edit-post/*", "/new-post").hasRole(ADMIN.name())
+                .anyRequest()
+                .authenticated()
+                .and()
+                .formLogin()
+                .loginPage("/login").permitAll();
+    }
+
+    @Override
+    @Bean
+    protected UserDetailsService userDetailsService() {
+        UserDetails adminUser = User.builder()
+                .username("steve")
+                .password(passwordEncoder.encode("password"))
+                .roles(ADMIN.name())
+                .build();
+
+        return new InMemoryUserDetailsManager(adminUser);
+    }
+}
