+++ src/main/java/com/stevedevblog/app/security/UserPermission.java
@@ -1,0 +1,16 @@
+package com.stevedevblog.app.security;
+
+public enum UserPermission {
+    POST_READ("read"),
+    POST_WRITE("write");
+
+    private final String permission;
+
+    UserPermission(String permission) {
+        this.permission = permission;
+    }
+
+    public String getPermission() {
+        return permission;
+    }
+}
