+++ src/main/java/com/stevedevblog/app/service/MarkdownRenderer.java
@@ -1,0 +1,31 @@
+package com.stevedevblog.app.service;
+
+import org.commonmark.Extension;
+import org.commonmark.ext.gfm.tables.TablesExtension;
+import org.commonmark.ext.image.attributes.ImageAttributesExtension;
+import org.commonmark.node.Node;
+import org.commonmark.parser.Parser;
+import org.commonmark.renderer.html.HtmlRenderer;
+import org.springframework.stereotype.Component;
+
+import java.util.List;
+
+@Component
+public class MarkdownRenderer {
+
+    private final Parser parser;
+    private final HtmlRenderer renderer;
+
+    public MarkdownRenderer() {
+        List<Extension> extensions = List.of(TablesExtension.create(), ImageAttributesExtension.create());
+        Parser parser = Parser.builder().extensions(extensions).build();
+        HtmlRenderer renderer = HtmlRenderer.builder().extensions(extensions).build();
+        this.parser = parser;
+        this.renderer = renderer;
+    }
+
+    public String markdownToHTML(String text) {
+        Node document = parser.parse(text);
+        return renderer.render(document);
+    }
+}
