--- src/main/java/com/stevedevblog/mvp/DevBlogApplication.java
@@ -1,13 +1,0 @@
-package com.stevedevblog.mvp;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DevBlogApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(DevBlogApplication.class, args);
-	}
-
-}
