--- src/main/java/com/stevedevblog/mvp/domain/PostCategory.java
@@ -1,18 +1,0 @@
-package com.stevedevblog.mvp.domain;
-
-public enum PostCategory {
-    BUILD_LOG("Build Log"),
-    TESTING("Testing"),
-    DESIGN_PATTERNS("Design Patterns");
-
-    private final String value;
-
-    PostCategory(String value) {
-
-        this.value = value;
-    }
-
-    public String getValue() {
-        return value;
-    }
-}
