--- src/main/java/com/stevedevblog/mvp/security/UserPermission.java
@@ -1,16 +1,0 @@
-package com.stevedevblog.mvp.security;
-
-public enum UserPermission {
-    POST_READ("read"),
-    POST_WRITE("write");
-
-    private final String permission;
-
-    UserPermission(String permission) {
-        this.permission = permission;
-    }
-
-    public String getPermission() {
-        return permission;
-    }
-}
