+++ src/main/resources/static/css/styles.css
@@ -38,6 +38,7 @@ h1, h2, h3 {
 }
 
 .nav-search-box {
+    outline: none;
     margin: 5px 0px;
 }
 
@@ -99,7 +100,12 @@ h1, h2, h3 {
 }
 
 .homepage-welcome {
+    margin-bottom: 20px;
+}
 
+.homepage-subheading {
+    margin-top: 20px;
+    margin-bottom:20px;
 }
 
 .welcome-title {
@@ -112,25 +118,54 @@ h1, h2, h3 {
     justify-content: flex-start;
 }
 
+.post-link {
+    color: rgb(0, 0, 0);
+    text-decoration: none;
+}
+
 .post-card {
-    border-top: 3px black solid;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    border: 1px rgb(158, 158, 158) solid;
     border-bottom: 3px black solid;
     width: 30%;
-    margin-bottom: 10px;
-    margin-left: 15px;
-    margin-right: 15px;
-    padding-bottom: 10px;
-    background-color: #929cad;
+    margin-bottom: 20px;
+    margin-left: 10px;
+    margin-right: 10px;
+    background-color: #e4e4e494;
     box-shadow: 0px 5px 5px #ccc;
-    }
+    transition: 1s;
+}
 
-.post-image {
-    width: 100%;
+.post-card:hover {
+    box-shadow: 0px 5px 5px black;
+}
+
+.listed-post-title {
+    margin-top: 5px;
+}
+
+.card-content-container {
+    display: flex;
+    flex-direction: column;
+    padding: 5px;
 }
 
 .card-text-content {
     padding: 5px;
 }
+
+.card-info-content {
+    padding: 10px;
+    text-align: right;
+}
+
+.post-image {
+    width: 100%;
+}
+
+
 
 /* new post */
 
@@ -235,27 +270,64 @@ h1, h2, h3 {
     background-color: #FFFFFF;
     border: 2px solid black;
     border-radius: 5px;
+    transition: 2s;
 }
 
 .login-button:hover {
     border: 2px solid #ed143d;
 }
 
-.read-more-button {
-    display: inline-block;
-    text-align: center;
-    width: 100%;
-    text-decoration: none;
-    color: #4b4f4f;
-    padding: 10px 0px;
-    background-color: #ccc;
-    transition: 0.3s;
-    border-top: 2px #ccc solid;
-    border-bottom: 2px #ccc solid;
+/* Small devices (landscape phones, 576px and up) */
+@media (max-width: 576px) {
+     body {
+         width: 100%;
+     }
+
+     .post-card {
+         width: 100%;
+         }
+
+     .nav-search {
+        display:none;
+     }
+
+     .nav-container {
+        padding-left: 20px;
+        display: flex;
+        flex-direction: column;
+     }
+
+}
+/* Medium devices (tablets, 768px and up) */
+@media (max-width: 768px) {
+    body {
+     width: 100%;
+     }
+
+    .post-card {
+      width: 100%;
+    }
+
+    .nav-container {
+        padding-left: 20px;
+        display: flex;
+        flex-direction: column;
+    }
+
+
 }
 
-.read-more-button:hover {
-     border-top: 2px #787878 solid;
-     border-bottom: 2px #787878 solid;
+/* Large devices (desktops, 992px and up) */
+@media (max-width: 992px) {
+
 }
 
+/* X-Large devices (large desktops, 1200px and up) */
+@media (max-width: 1200px) {
+
+}
+
+/* XX-Large devices (larger desktops, 1400px and up) */
+@media (max-width: 1400px) {
+
+}
