+++ src/main/resources/templates/partials/header.mustache
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1" />
     <link rel="stylesheet" href="/css/styles.css" />
     <link rel="stylesheet" href="/css/post-markdown-styles.css" />
     <link rel="icon" href="/icons/favicon.svg">
