+++ src/main/resources/templates/partials/navbar.mustache
@@ -2,7 +2,7 @@
     <div>
         <a class="site-title" href="/">SteveDevBlog</a>
     </div>
-    <nav>
+    <nav class="nav-container">
         <a class="nav-link" href="/blogs">Blogs</a>
         <a class="nav-link" href="/newsletter">Newsletter</a>
         <a class="nav-link" href="/about">About</a>
